Fixing tests broken by new invoice functionality.
Outdated version of FactoryGirl made this a true PITA. Had to manually
create join table associations. UGH.

[53918671]

+++ db/migrate/20130730232020_create_invoice_items.rb
@@ -2,9 +2,9 @@ class CreateInvoiceItems < ActiveRecord::Migration
   def change
     create_table :invoice_items do |t|
       t.string :name
-      t.decimal :amount
-      t.decimal :hours
-      t.decimal :total
+      t.float :amount
+      t.float :hours
+      t.float :total
 
       t.references :invoice
